requirements.txt
python-telegram-bot>=20.7
aiohttp>=3.9.0
aiosqlite>=0.19.0
Pillow>=10.0.0
python-dotenv>=1.0.0